Emergency Call when there is no room left for new Call.
Initiate Emergency Call when there is already active
and held calls are present in the Call list. The behavior
would be the Current active call to be Disconnected and
Emergency Call to be initiated in place of active call.
Functionality to disconnect the current active call is
already existing in the code but does not work. The reason
is When Emergency Call is initiated immediately after the
current call is disconnected, It gets failed from Telephony
frameworks as the Disconnection of the active call is not
yet notified from lower layers and frameworks still has got
the Active call in the list.
With this change, Emergency call would be kept in waiting
until Callsmanager gets notified Disconnection status of
Current active call and initiate the Emergency Call now.
Also Reject Incoming or waiting Call while initiating an
Emergency Call.

Change-Id: If4e82eb3d9343d677be496afd1b2e2c9b5a88b86
CRs-Fixed: 2205288 2208253