IMS: Fix swap issue for active call with conference call
Conference child calls are treated as normal calls and are
being returned when queried for current focus call.
When active call is swapped with conference call, Telephony
is trying to act on child call which is causing the issue.

To fix this, do not update child call as current focus call.

Change-Id: Ice4769edbb776889ae738cfa850625ca0a548600
CRs-Fixed: 2202634