Enable "Respond via SMS" option.
In current implementation, when "Ask First" is enabled for SMS
preferred setting, then "Respond via SMS" option is shown for
MT call on one sub and not shown for MT call on other sub leading
to inconsistency.
As part of this fix, enable "Respond via SMS" option for both subs
regardless of what is SMS preferred setting.

Change-Id: I62108309d06ea49659a09c7ec15c551b6d4126b6
CRs-Fixed: 2199011